requirements.txt
asgiref==3.4.1
black==21.12b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Django==4.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==4.0.1
tzdata==2021.5
